trying to integrate gradient descent and adam with the neural network itself

+++ r_nn/src/nn/node/mod.rs
@@ -4,14 +4,14 @@ pub mod node {
     use std::fmt;
 
     /// calculates the dot product of two ndarray type objects
-    pub fn dot_product(arr1: ArrayView1<f32>, arr2: ArrayView1<f32>) -> f32 {
+    pub fn dot_product(arr1: ArrayView1<f64>, arr2: ArrayView1<f64>) -> f64 {
         arr1.dot(&arr2)
     }
 
     #[derive(Clone)]
     pub struct Node {
-        pub weights: Vec<f32>,
-        pub bias: f32
+        pub weights: Vec<f64>,
+        pub bias: f64
     }
 
     impl fmt::Display for Node {
@@ -22,28 +22,28 @@ pub mod node {
 
     impl Node {
         /// Creates a new Node Object with weights and a bias
-        pub fn new(weights: Vec<f32>, bias: f32) -> Self {
+        pub fn new(weights: Vec<f64>, bias: f64) -> Self {
             Node {
                 weights,
                 bias
             }
         }
-        pub fn sigmoid_actualize(&self, features: Vec<f32>) -> f32 {
-            let features_arr: ndarray::prelude::ArrayBase<ndarray::OwnedRepr<f32>, ndarray::prelude::Dim<[usize; 1]>> = Array1::from_vec(features);
+        pub fn sigmoid_actualize(&self, features: Vec<f64>) -> f64 {
+            let features_arr: ndarray::prelude::ArrayBase<ndarray::OwnedRepr<f64>, ndarray::prelude::Dim<[usize; 1]>> = Array1::from_vec(features);
             let weights_arr: ndarray::prelude::ArrayBase<ndarray::OwnedRepr<_>, ndarray::prelude::Dim<[usize; 1]>> = Array1::from_vec(self.weights.clone());
             println!("calculating dot product of features and weights");
             println!("Features: {}", features_arr.view());
             println!("Weights: {}", weights_arr.view());
-            let z: f32 = dot_product(features_arr.view(), weights_arr.view()) + self.bias;
+            let z: f64 = dot_product(features_arr.view(), weights_arr.view()) + self.bias;
             println!("done");
-            let exp_z: f32 = (-z).exp();
+            let exp_z: f64 = (-z).exp();
             1.0 / (1.0 + exp_z)
         }
-        pub fn set_weights(&mut self, weights: Vec<f32>, bias: f32) {
+        pub fn set_weights(&mut self, weights: Vec<f64>, bias: f64) {
             self.weights = weights;
             self.bias = bias;
         }
-        pub fn get_weights(&self) -> (Vec<f32>, f32) {
+        pub fn get_weights(&self) -> (Vec<f64>, f64) {
             (self.weights.clone(), self.bias)
         }
         pub fn str(&self) {
